update by ochim

+++ JasaService/app/src/main/java/com/rohim/fragment/FragmentRequestOrderActive.java
@@ -5,12 +5,15 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
 import com.rohim.adapter.RecycleViewListAdapterDetailRequest;
 import com.rohim.common.BaseFragment;
+import com.rohim.common.PopupNotification;
+import com.rohim.enumeration.EnumInputService;
 import com.rohim.jasaservice.MainActivity;
 import com.rohim.jasaservice.R;
 import com.rohim.modal.RequestDetail;
@@ -33,6 +36,7 @@ public class FragmentRequestOrderActive extends BaseFragment {
     private RecycleViewListAdapterDetailRequest adapterListView;
     Button btnBack, btnCancel;
     TextView textTitle, textUserName, textNoTelp;
+    String idRequest, idUserAccepted, userNameAccepted, userNoTelp, msgId, msgTitle, msgBody;
 
     @Override
     public void initView() {
@@ -41,22 +45,71 @@ public class FragmentRequestOrderActive extends BaseFragment {
         textTitle = (TextView) view.findViewById(R.id.text_title_request_order);
         textUserName = (TextView) view.findViewById(R.id.text_user_name_request_order);
         textNoTelp = (TextView) view.findViewById(R.id.text_no_telp_request_order);
-        loadInit();
-        textUserName.setText("Ahmad Sobirin \nProfesional "+textTitle.getText().toString());
-        textNoTelp.setText("No. Telp : 0856752318998");
-
         btnCancel = (Button) view.findViewById(R.id.btn_cancel_request_order);
+        btnBack = (Button) view.findViewById(R.id.btn_back_request_order);
+        if(msgId!=null){
+            notifMode();
+        }
+        loadInit();
+//        textUserName.setText("Ahmad Sobirin \nProfesional "+textTitle.getText().toString());
+//        textNoTelp.setText("No. Telp : 0856752318998");
+
+
         btnCancel.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                db = dbh.getWritableDatabase();
-                String sql = "DELETE FROM "+RequestOrder.tbl_request_order;
-                db.execSQL(sql);
-                sql = "DELETE FROM "+RequestDetail.tbl_request_detail;
-                db.execSQL(sql);
-                db.close();
+                if(btnCancel.getText().toString().equalsIgnoreCase("Cancel")){
+                    // do synchronize server
+
+                    // do delete data from device
+                    db = dbh.getWritableDatabase();
+                    String sql = "DELETE FROM "+RequestOrder.tbl_request_order;
+                    db.execSQL(sql);
+                    sql = "DELETE FROM "+RequestDetail.tbl_request_detail;
+                    db.execSQL(sql);
+                    db.close();
+                }else{
+                    // do submit comment
+
+                }
+
             }
         });
+
+        btnBack.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                getFragmentManager().popBackStack();
+            }
+        });
+    }
+
+    private void notifMode() {
+        openDatabaseHelper();
+        try {
+            if(msgId.equals("PROCESS") || msgId.equals("CANCEL1")) {
+                RequestOrder ro = requestOrderDao.queryForId(idRequest);
+                ro.setStatus(msgId);
+                ro.setUserName(userNameAccepted);
+                ro.setUserNoTelfon(userNoTelp);
+                requestOrderDao.update(ro);
+
+            }
+            if(msgId.equals("FINISH")){
+                RequestOrder ro = requestOrderDao.queryForId(idRequest);
+                ro.setStatus(msgId);
+                requestOrderDao.update(ro);
+
+            }
+            PopupNotification popupNotification = new PopupNotification();
+            popupNotification.setParam(getContext(), msgTitle, msgBody, false);
+            popupNotification.show(getFragmentManager(), "");
+
+        } catch (SQLException e) {
+            Log.e("Error", e.toString());
+        } finally {
+            dbh.close();
+        }
     }
 
     private void loadInit() {
@@ -78,10 +131,12 @@ public class FragmentRequestOrderActive extends BaseFragment {
                         +" FROM "+ RequestOrder.tbl_request_order +" A LEFT JOIN "+ RequestDetail.tbl_request_detail+" B "
                         +" ON A."+RequestOrder.clm_id_request+" = B."+RequestDetail.clm_fid_request
                         +" LEFT JOIN "+ Service.tbl_service+" C ON A.FID_SERVICE = C.ID_SERVICE"
-                        +" AND A."+RequestOrder.clm_status+" IN ('NEW', 'ACCEPT') ORDER BY B."+RequestDetail.clm_id_request_detail+" ASC";
+                        +" AND A."+RequestOrder.clm_status+" IN ('NEW', 'PROCESS', 'CANCEL1') ORDER BY B."+RequestDetail.clm_id_request_detail+" ASC";
                 // get db conection
                 db = dbh.getWritableDatabase();
                 cursor = db.rawQuery(sql, null);
+
+
                 if (cursor.moveToFirst()) {
                     do {
                         RequestDetail rd = new RequestDetail();
@@ -99,10 +154,42 @@ public class FragmentRequestOrderActive extends BaseFragment {
                     } while (cursor.moveToNext());
                 }
             } catch (Exception e) {
-                e.printStackTrace();
+                Log.e("Error", e.toString());
             }finally {
                 cursor.close();
                 db.close();
+            }
+        }
+
+        if(msgId!=null && msgId.equals("FINISH")){
+            openDatabaseHelper();
+            try {
+                RequestOrder ro = requestOrderDao.queryForId(idRequest);
+                Service service = serviceDao.queryForId(ro.getFidService());
+                textTitle.setText(service.getServiceName());
+                textUserName.setText(ro.getUserName());
+                textNoTelp.setText(ro.getUserNoTelfon());
+
+                data.clear();
+                RequestDetail rd1 = new RequestDetail();
+                rd1.setServiceItemName("Hasil Service");
+                rd1.setSatuan("REASON_COMENT");
+                rd1.setJenisInput(EnumInputService.SpinnerInput.getVal());
+
+                RequestDetail rd2 = new RequestDetail();
+                rd2.setServiceItemName("Ulasan");
+                rd2.setJenisInput(EnumInputService.TextLong.getVal());
+
+                data.add(rd1);
+                data.add(rd2);
+                adapterListView = new RecycleViewListAdapterDetailRequest(getContext(), data, fragmentManager, false);
+
+                btnCancel.setText("SUBMIT");
+                btnBack.setVisibility(View.INVISIBLE);
+            } catch (SQLException e) {
+                Log.e("Error", e.toString());
+            } finally {
+                dbh.close();
             }
         }
     }
@@ -113,5 +200,62 @@ public class FragmentRequestOrderActive extends BaseFragment {
         listview.setLayoutManager(new LinearLayoutManager(getContext()));
         adapterListView = new RecycleViewListAdapterDetailRequest(getContext(), data, fragmentManager, true);
         listview.setAdapter(adapterListView);
+    }
+
+    public String getIdRequest() {
+        return idRequest;
+    }
+
+    public void setIdRequest(String idRequest) {
+        this.idRequest = idRequest;
+    }
+
+    public String getIdUserAccepted() {
+        return idUserAccepted;
+    }
+
+    public void setIdUserAccepted(String idUserAccepted) {
+        this.idUserAccepted = idUserAccepted;
+    }
+
+    public String getUserNameAccepted() {
+        return userNameAccepted;
+    }
+
+    public void setUserNameAccepted(String userNameAccepted) {
+        this.userNameAccepted = userNameAccepted;
+    }
+
+    public String getUserNoTelp() {
+        return userNoTelp;
+    }
+
+    public void setUserNoTelp(String userNoTelp) {
+        this.userNoTelp = userNoTelp;
+    }
+
+    public String getMsgId() {
+        return msgId;
+    }
+
+    public void setMsgId(String msgId) {
+        this.msgId = msgId;
+    }
+
+
+    public String getMsgTitle() {
+        return msgTitle;
+    }
+
+    public void setMsgTitle(String msgTitle) {
+        this.msgTitle = msgTitle;
+    }
+
+    public String getMsgBody() {
+        return msgBody;
+    }
+
+    public void setMsgBody(String msgBody) {
+        this.msgBody = msgBody;
     }
 }
